Testing library linking issues fixes. Tests fixes.

+++ include/nil/actor/core/iostream-impl.hh
@@ -279,7 +279,7 @@ namespace nil {
 
         template<typename CharType>
         future<> input_stream<CharType>::skip(uint64_t n) {
-            auto skip_buf = std::min(n, _buf.size());
+            auto skip_buf = std::min(static_cast<std::size_t>(n), _buf.size());
             _buf.trim_front(skip_buf);
             n -= skip_buf;
             if (!n) {
